binary search

+++ Algorithms/Search/Linear Search/linear_search.cpp
@@ -4,8 +4,9 @@ using namespace std;
 
 /**
  * searches a element in a given vector
+ * returns the index of the element, if it exists, or -1, otherwise
 */
-long long int linear_search(vector<long long int> list, long long int element) {
+int linear_search(vector<int> list, int element) {
     // iterates over the vector in order
     for(int i = 0; i < list.size(); i++){
         if(list[i] == element)
